proyecto/requirements.txt
# psycopg2-binary==2.9.9
# pandas==2.1.4
# openpyxl==3.1.2
# python-dotenv==1.0.0

psycopg2-binary
pandas
openpyxl
python-dotenv
boto3
sqlalchemy
